requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
pymongo==4.8.0
plotly==5.24.1
streamlit==1.40.1
python-dotenv==1.0.1
joblib==1.4.2
